Add UpdateFastlaneCIService class
This PR is ready to be merged - however UpdateFastlaneCIService needs some more updating as soon as we decided on how exactly we're gonna distribute fastlane.ci via RubyGems. In particular how to launch up the server is going to be relevant

As part of UpdateFastlaneCIService, we'll also need some code to check if any background workers or build workers are still running. I left the code empty for now, as we haven't yet finished our build workers setup, and we'd end up re-implementing things if we wrote it today